Rewrite the contributing guide
The goal behind this document should be to explain to new and potential
contributors what they need to do (and follow) to get involved and it
wasn't really doing that.

It should also be a document that we can link to when closing or explaining
reasoning behind decisions. Broadly, it should aim to ensure decisions
aren't just arbitary.